rolled back from half-finished offset transformation changes.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@2042 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcaqgui2d/iknowsplatformposition2d.h
@@ -24,10 +24,9 @@ public:
     // all purport to know the position.
     virtual int platformKnowledgeReliability() const=0;
 
-    // coordinates in GUI CS
-    virtual double x() const=0;
-    virtual double y() const=0;
-    virtual double theta() const=0;
+    virtual float x() const=0;
+    virtual float y() const=0;
+    virtual float theta() const=0;
 
 private:
